pyproject.toml
[tool.poetry]
name = "review_etl"
version = "0.1.0"
description = "Extraction of Google Play Store reviews"
authors = ["Ibrahim Cikrikcioglu <ibrahimcikrikcioglu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
pymango = "^0.1.1"
google-play-scraper = "^1.2.3"
pyyaml = "^6.0"
pymongo = "^4.3.3"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
pytest = "^7.2.1"
black = "^23.1.0"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true
